+++ src/main/kotlin/io/kjson/mustache/TemplatePartial.kt
@@ -2,7 +2,7 @@
  * @(#) TemplatePartial.kt
  *
  * mustache-k  Mustache template processor for Kotlin
- * Copyright (c) 2020, 2021 Peter Wall
+ * Copyright (c) 2020, 2021, 2023 Peter Wall
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -25,6 +25,8 @@
 
 package io.kjson.mustache
 
+import net.pwall.util.CoOutput
+
 /**
  * A Template Partial that refers to another Template.
  *
@@ -36,6 +38,10 @@ class TemplatePartial : Partial {
 
     override fun appendTo(appendable: Appendable, context: Context) {
         template.appendTo(appendable, context)
+    }
+
+    override suspend fun outputTo(out: CoOutput, context: Context) {
+        template.outputTo(out, context)
     }
 
 }