+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ## [Unreleased]
 ### Changed
 - `pom.xml`: updated dependencies, updated Kotlin version to 1.7.21
+- `Section`, `InvertedSection`: added coverage of unsigned numbers
 
 ## [1.3] - 2022-01-28
 ### Changed